+++ src/jsMain/kotlin/at/posselt/pfrpg2e/actor/Game.kt
@@ -1,9 +1,7 @@
 package at.posselt.pfrpg2e.actor
 
 import com.foundryvtt.core.Game
-import com.foundryvtt.pf2e.actor.PF2EArmy
 import com.foundryvtt.pf2e.actor.PF2ECharacter
-import com.foundryvtt.pf2e.actor.PF2EFamiliar
 import com.foundryvtt.pf2e.actor.PF2ENpc
 import com.foundryvtt.pf2e.actor.PF2EParty
 import js.array.toTypedArray
@@ -21,18 +19,4 @@ fun Game.npcs(): Array<PF2ENpc> =
     actors.contents
         .asSequence()
         .filterIsInstance<PF2ENpc>()
-        .toTypedArray()
-
-fun Game.playerFamiliars(): Array<PF2EFamiliar> =
-    actors.contents
-        .asSequence()
-        .filterIsInstance<PF2EFamiliar>()
-        .filter { it.hasPlayerOwner }
-        .toTypedArray()
-
-fun Game.playerArmies(): Array<PF2EArmy> =
-    actors.contents
-        .asSequence()
-        .filterIsInstance<PF2EArmy>()
-        .filter { it.hasPlayerOwner }
         .toTypedArray()