+++ src/jsMain/kotlin/at/posselt/pfrpg2e/kingdom/Kingdom.kt
@@ -3,9 +3,9 @@ package at.posselt.pfrpg2e.kingdom
 import at.posselt.pfrpg2e.utils.getAppFlag
 import at.posselt.pfrpg2e.utils.setAppFlag
 import com.foundryvtt.core.utils.deepClone
-import com.foundryvtt.pf2e.actor.PF2ENpc
+import com.foundryvtt.pf2e.actor.PF2EParty
 
-typealias KingdomActor = PF2ENpc
+typealias KingdomActor = PF2EParty
 
 fun KingdomActor.getKingdom(): KingdomData? =
     getAppFlag<KingdomActor, KingdomData?>("kingdom-sheet")