+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Structure Token Mapping Macro
 * Removed Roll Kingdom Event macro since those depend on the actual kingdom sheet now
+* Removed Camping and Kingdom macros; those are now accessed through the party actor
 
 ### Changed
 * Armies don't need to be targeted anymore but simply selected